% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastymd-package.R
\docType{package}
\name{fastymd-package}
\alias{fastymd}
\alias{fastymd-package}
\title{fastymd: Fast Utilities for Year Month Day Objects}
\description{
A collection of utility functions for working with Year Month Day objects. Includes functions for fast parsing of numeric and character input based on algorithms described in Hinnant, H. (2021) \url{https://howardhinnant.github.io/date_algorithms.html} as well as a branchless calculation of leap years by Jerichaux (2025) \url{https://stackoverflow.com/a/79564914}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://timtaylor.codeberg.page/fastymd/}
  \item Report bugs at \url{https://codeberg.org/TimTaylor/fastymd/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})

Authors:
\itemize{
  \item Howard Hinnant (Author of underlying algorithms for calculating Epoch days to Calendar days and vice versa)
  \item jerichaux (Author of included branchless leap year calculation)
}

}
\keyword{internal}
