% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\alias{is_leap}
\title{Is value a leap year?}
\usage{
is_leap_year(x)

is_leap(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
\code{logical} result.
}
\description{
Determine whether an input value is a leap year using the branchless approach
of jerichaux (2025). Method provided for both Dates and numeric values.
Numeric values first floored before the calculation is made.
}
\references{
jerichaux. (2025) \emph{How to find leap year programmatically in C}.
Available at: \url{https://stackoverflow.com/a/79564914}
(Accessed 16 April 2025).
}
